+++ src/App.css
@@ -1,7 +1,3 @@
 #main {
   margin: 0 10%;
 }
-
-.body {
-  text-align: center;
-}
