+++ src/containers/Caregroup/Caregroup.js
@@ -1,18 +1,54 @@
 import React, { Component } from 'react';
+import './Caregroup.css';
+import Config from './../../config.json';
+import Axios from 'axios';
 import CaregroupCard from '../../components/Card/CaregroupCard';
 
-import './Caregroup.css';
+class CareGroup extends Component {
 
-export default class CareGroup extends Component {
+  constructor() {
+    super();
+    this.state = {
+      caregroups: [],
+      result: {}
+    };
+  }
+
+  componentDidMount() {
+    const config = {
+      auth: {
+        username: process.env.REACT_APP_API_USER,
+        password: process.env.REACT_APP_API_PW
+      }
+    };
+    Axios.get(Config.api.host + '/care-groups', config).then(res => {
+      this.setState({
+        caregroups: res.data
+      });
+    });
+  }
+
+  /**
+   * Populate card with data from the api
+   */
+  populate = () => {
+    if (this.state.caregroups.length === 0) {
+      return null;
+    }
+    return this.state.caregroups.map((cg, index) => {
+      return (
+        <CaregroupCard caregroup={cg} key={index}/> 
+      );
+    });
+  };
+
   render() {
     return (
       <React.Fragment>
         <h1>Care Group</h1>
-        <div className='caregroup'>
-          <CaregroupCard />
-        </div>
-
+        {this.populate()}
       </React.Fragment>
     );
   }
 }
+export default CareGroup;
